pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.8",
    "langchain>=0.3.19",
    "langchain-community>=0.3.18",
    "langchain-openai>=0.3.7",
    "openai>=1.65.2",
    "pymssql>=2.3.2",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.38",
    "streamlit>=1.42.2",
    "uvicorn>=0.34.0",
]
